src/requirements.txt
hypercorn
fastapi
pandas
httpx
python-dotenv